Fix bugs

+++ label_studio/projects/api.py
@@ -198,7 +198,7 @@ class ProjectAPI(APIViewVirtualRedirectMixin,
         label_config = self.request.query_params.get('label_config')
 
         # config changes can break view, so we need to reset them
-        if parse_config(label_config) != parse_config(project.label_config):
+        if label_config and parse_config(label_config) != parse_config(project.label_config):
             View.objects.filter(project=project).all().delete()
 
         return super(ProjectAPI, self).patch(request, *args, **kwargs)
